requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==3.2
django-allauth==0.51.0
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
executing==0.8.3
fonttools==4.33.3
idna==3.3
ipython==8.3.0
jedi==0.18.1
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.0
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.9.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyodbc==4.0.32
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tenacity==8.0.1
traitlets==5.2.1.post0
urllib3==1.26.11
wcwidth==0.2.5
